Split up tests
The test file was getting cluttered, so I split the tests up into
different files. This entailed

1) adding a test/ dir to keep things tidy,
2) renaming tests and test functions, and
3) reworking the makefile.

The makefile isn't perfect -- it seems too abstract in some places,
and at the same time not abstract enough in others -- but it was
surprisingly enjoyable to work on. Some think that time spent on
makefiles is time wasted, but I only look at them a few times a year,
so why not have some fun with it?

I also went ahead and added commentary to each individual test file.
This should perhaps be added in a separate commit, since it's new
stuff, but...who cares?