Add extended database test
A Prolog-style query language is developed in SICP seciton 4.4. Also
provided for sake of example is "the data base of personnel records
for Microshaft, a thriving high-technology company in the Boston
area". The database and the queries from that section are included
here, showing that Reazon can be used with a database and also that
miniKanren could replace Prolog in SICP.