Unroll Sudoku subseto calls
The Sudoku solvers were using the general `subseto` relation, but in
fact that generality is not needed: the groups of squares are
stipulated to contain only very specific items, and some speed can be
gained by taking advantage of this. A little elegance is thereby lost,
but oh well.

Running the 4x4 benchmark on a 2012 MBP:

Before:

Elapsed time: 31.900794s (20.630957s in 172 GCs)
Elapsed time: 31.963738s (20.649389s in 171 GCs)

After:

Elapsed time: 25.398572s (16.814765s in 143 GCs)
Elapsed time: 25.441664s (16.857431s in 142 GCs)