Precompute 4x4 Sudoku permutations
Previously, the 4x4 solver was effectively computing the permutations
of (1 2 3 4) over and over at runtime. A huge performance boost can be
gotten by computing the permutations once at compile time, then
running a disjunction of == calls to each of the permutations with
each of the blocks.

The resulting macro expansion is about 300 lines, and this is what it
looks like:

(reazon-run* (a1 a2 a3 a4 b1 b2 b3 b4 c1 c2 c3 c4 d1 d2 d3 d4)
  (reazon-disj
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (1 2 3 4)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (1 2 4 3)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (1 4 2 3)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (1 3 2 4)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (1 4 3 2)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (1 3 4 2)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (2 1 3 4)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (2 1 4 3)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (4 1 2 3)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (2 4 1 3)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (3 1 2 4)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (2 3 1 4)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (4 1 3 2)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (4 2 1 3)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (2 4 3 1)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (3 1 4 2)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (2 3 4 1)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (3 2 1 4)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (4 3 1 2)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (4 2 3 1)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (3 4 1 2)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (3 2 4 1)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (4 3 2 1)))
   (reazon-== (\` ((\, a1) (\, a2) (\, a3) (\, a4))) (quote (3 4 2 1))))
...
)

Unfortunately, this strategy doesn't seem to scale to the 9x9 case.
This is because there are just 24 permutations of a four-element list,
but 362880 for a nine-element list.

Running the 4x4 Sudoku benchmark on a 2012 MBP (where the Before time
uses the previous unrolling strategy):

Before:

Elapsed time: 24.264427s (16.165260s in 212 GCs)
Elapsed time: 24.282203s (16.141869s in 210 GCs)

After:

Elapsed time: 5.426367s (3.888877s in 58 GCs)
Elapsed time: 5.307984s (3.790005s in 57 GCs)

Note that we've gone from spending around 2/3 of time in GC to almost
3/4. Why is that?